test(unit): conver remaining tests to vitest
closes #458
