+++ .eslintrc.yml
@@ -2,4 +2,3 @@ root: true
 extends:
   - '@form8ion'
   - '@form8ion/cucumber'
-  - '@form8ion/mocha'
