+++ package.json
@@ -44,9 +44,7 @@
     "build": "npm-run-all --print-label --parallel build:*",
     "prepack": "run-s build",
     "test:unit": "cross-env NODE_ENV=test c8 run-s test:unit:base",
-    "test:unit:base": "npm-run-all --print-label --parallel test:unit:mocha test:unit:vitest",
-    "test:unit:mocha": "DEBUG=any mocha 'src/**/*-test.js'",
-    "test:unit:vitest": "DEBUG=any vitest run",
+    "test:unit:base": "DEBUG=any vitest run",
     "lint:peer": "npm ls >/dev/null",
     "lint:engines": "ls-engines",
     "lint:publish": "publint --strict"
@@ -74,7 +72,6 @@
     "@form8ion/commitlint-config": "2.0.4",
     "@form8ion/eslint-config": "7.0.12",
     "@form8ion/eslint-config-cucumber": "1.4.1",
-    "@form8ion/eslint-config-mocha": "3.0.6",
     "@form8ion/remark-lint-preset": "6.0.6",
     "@rollup/plugin-babel": "6.0.4",
     "@rollup/plugin-node-resolve": "16.0.0",
@@ -91,7 +88,6 @@
     "js-yaml": "4.1.0",
     "lockfile-lint": "4.14.0",
     "ls-engines": "0.9.3",
-    "mocha": "11.1.0",
     "mock-fs": "5.4.1",
     "npm-run-all2": "7.0.2",
     "publint": "0.3.2",
@@ -101,7 +97,6 @@
     "rimraf": "6.0.1",
     "rollup": "4.34.0",
     "rollup-plugin-auto-external": "2.0.0",
-    "sinon": "19.0.2",
     "vitest": "3.0.4"
   }
 }
